Split KFFLang and KFFLib into two Gradle projects
