+++ gradle.properties
@@ -4,13 +4,14 @@ kotlin.code.style=official
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
+# https://github.com/JetBrains/kotlin
 kotlin_version=1.7.10
 max_kotlin=1.8
-
+# https://github.com/Kotlin/kotlinx.coroutines
 coroutines_version = 1.6.3
 max_coroutines=1.7
 
 serialization_version = 1.3.3
 max_serialization=1.4
-
+# This needs to match whatever Forge is using in dev
 annotations_version = 23.0.0
