+++ kfflang/src/main/kotlin/thedarkcolour/kotlinforforge/forge/KDeferredRegister.kt
@@ -3,9 +3,6 @@
 package thedarkcolour.kotlinforforge.forge
 
 import net.minecraftforge.registries.DeferredRegister
-import java.util.function.Supplier
-import kotlin.properties.ReadOnlyProperty
-import kotlin.reflect.KProperty
 
 /**
  * Got rid of KDeferredRegister because KFF can no longer access forge/mc code
@@ -31,18 +28,18 @@ public typealias KDeferredRegister<T> = DeferredRegister<T>
  * @param V Type of deferred register
  * @param T Specific type of object being registered
  */
-public inline fun <V, T : V> KDeferredRegister<V>.registerObject(
-    name: String,
-    noinline supplier: () -> T
-): ReadOnlyProperty<Any?, T> {
-    val registryObject = this.register(name, supplier)
-
-    // note that this anonymous class inherits three types
-    return object : ReadOnlyProperty<Any?, T>, Supplier<T>, () -> T {
-        override fun get(): T = registryObject.get()
-
-        override fun getValue(thisRef: Any?, property: KProperty<*>): T = get()
-
-        override fun invoke(): T = get()
-    }
-}
+//public inline fun <V, T : V> KDeferredRegister<V>.registerObject(
+//    name: String,
+//    noinline supplier: () -> T
+//): ReadOnlyProperty<Any?, T> {
+//    val registryObject = this.register(name, supplier)
+//
+//    // note that this anonymous class inherits three types
+//    return object : ReadOnlyProperty<Any?, T>, Supplier<T>, () -> T {
+//        override fun get(): T = registryObject.get()
+//
+//        override fun getValue(thisRef: Any?, property: KProperty<*>): T = get()
+//
+//        override fun invoke(): T = get()
+//    }
+//}
