+++ kfflang/src/main/kotlin/thedarkcolour/kotlinforforge/package.md
@@ -1,0 +1,2 @@
+# thedarkcolour.kotlinforforge
+This package contains the language loader classes for discovering and initializing Kotlin mods
