+++ kfflang/src/main/resources/META-INF/MANIFEST.MF
