+++ kfflang/src/test/resources/META-INF/mods.toml
@@ -5,7 +5,7 @@ license="LGPL v2.1"
 [[mods]]
 displayName="Kotlin for Forge"
 modId="kotlinforforgetest"
-version="3.6.0"
+version="3.7.0"
 authors="TheDarkColour"
 credits="Herobrine knows all. Thanks to Username404-59 and TheOnlyTails on GitHub"
 issueTrackerURL="https://github.com/thedarkcolour/KotlinForForge/issues"
