+++ kfflang/src/test/resources/pack.mcmeta
