+++ kfflib/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+FMLModType: GAMELIBRARY
+Automatic-Module-Name: kfflib
