+++ settings.gradle.kts
@@ -9,3 +9,5 @@ pluginManagement {
         id("org.jetbrains.kotlin.plugin.serialization") version kotlin_version
     }
 }
+include("kfflib")
+include("kfflang")
